requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
fastapi
jinja2
uvicorn
python-multipart
ipykernel
python-dotenv
tavily-python
langchainhub
duckduckgo-search
wikipedia
numexpr